# app/docs/swagger.yaml
definitions:
...
  user.UserLoginReq:
    properties:
      account:
        description: 登录账号
        type: string
      area_code:
        description: 区号,如果是手机号 传 区号，如果是邮箱 传 邮箱服务商 gmail、qq、outlook 等
        type: string
      code:
        description: 验证码
        type: string
      identity_type:
        description: 登录类型，mobile_phone、email 目前只有 mobile_phone
        type: string
      login_type:
        description: 登录方式 password 密码登录、verify_code 验证码登录
        type: string
      password:
        description: |-
          登录密码，前端 必须进行 sha256 hash 后传过来 【小写】
          1. 如果账号存在的时候登录，密码 或者验证码必须有一个
          2. 如果获取 验证码登录的时候，获取验证码返回的是新用户，需要设置密码，也是使用该字段
        type: string
    required:
    - account
    - area_code
    - identity_type
    - login_type
    type: object
  user.UserLoginResp:
    properties:
      is_new:
        description: 是否是新用户
        type: boolean
      token:
        description: 用户登录成功的 token
        type: string
      user_sys_id:
        description: uid 用户id
        type: string
    type: object
info:
  contact: {}
paths:
...
  /user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户登录
        in: body
        name: LoginReq
        required: true
        schema:
          $ref: '#/definitions/user.UserLoginReq'
      - description: 请求ID
        in: header
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/user.UserLoginResp'
      summary: 手机\邮箱验证码登录
      tags:
      - user
...
swagger: "2.0"